Adjust spacing and sizes for different aggregation UI screens (loading, error, no-data, query error)
