--- client/web/src/search/results/components/aggregation/AggregationBarsBackground.module.scss
@@ -1,10 +1,0 @@
-.zero-state-background {
-    width: 100%;
-    margin: auto;
-
-    // Visually compensate bottom spacing between error chart mock
-    // UI and aggregation controls UI which are placed below this
-    // zero state block
-    padding-bottom: 1.5rem;
-    padding-top: 0.5rem;
-}
