+++ client/web/src/search/results/components/aggregation/components/aggregation-chart-card/AggregationChartCard.module.scss
@@ -1,10 +1,3 @@
-// Chart (parent-size) container
-.container {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-}
-
 .chart {
     flex-grow: 1;
 }
@@ -20,7 +13,7 @@
 
 // Internal error-alert UI
 .error-alert {
-    margin: 0 0 0.5rem 0;
+    margin: 0;
     overflow: auto;
 }
