+++ client/web/src/search/results/components/aggregation/components/aggregation-chart-card/components/index.ts
@@ -1,0 +1,6 @@
+export { AggregationTextContent, AggregationContent } from './AggregationLayouts'
+
+// Note that this is a type export, and we don't export AggregationChart
+// itself here. This is because we lazy load AggregationChart due to its size
+// you should export this component through lazy load helper directly into consumer
+export type { AggregationChartProps } from './AggregationChart'
