+++ client/web/src/search/results/components/aggregation/components/aggregation-mode-controls/AggregationModeControls.module.scss
