+++ client/web/src/search/results/components/aggregation/components/index.ts
@@ -1,0 +1,3 @@
+export { AggregationLimitLabel } from './AggregationLimitLabel'
+export { AggregationModeControls } from './aggregation-mode-controls/AggregationModeControls'
+export { AggregationChartCard } from './aggregation-chart-card/AggregationChartCard'
