+++ client/web/src/search/results/components/aggregation/index.ts
@@ -1,8 +1,8 @@
 export * from './types'
 export * from './hooks'
 
-export { AggregationChartCard, getAggregationData } from './AggregationChartCard'
-export { AggregationModeControls } from './AggregationModeControls'
+export { AggregationChartCard, getAggregationData } from './components/aggregation-chart-card/AggregationChartCard'
+export { AggregationModeControls } from './components/aggregation-mode-controls/AggregationModeControls'
 export { SearchAggregationResult } from './SearchAggregationResult'
-export { AggregationLimitLabel } from './AggregationLimitLabel'
+export { AggregationLimitLabel } from './components/AggregationLimitLabel'
 export { GroupResultsPing } from './pings'
